refactor(platform-api): DRY wrapper layer + fix audit ForbiddenError
- Audit wrapper threw `new Error('Forbidden')` instead of `ForbiddenError`,
  bypassing adapter's status mapping. Now throws `ForbiddenError` → 403.
- Drop unused `append()` from `AuthorizedAuditEventRepository` — engine
  uses raw repo (god mode); no handler calls it.
- Extract `filterByParentNamespace<T>` helper for indirect-namespace
  wrappers. Replaces 13-line copies in HumanTask + Handoff wrappers.
- Move `assertNamespaceWrite(ns)` into `AuthorizedRepository` base.
  Replaces 6 copy-pasted private helpers.
- Merge `RouteAdapterOptionsInternal` into `RouteAdapterOptions`.
- `AuthorizedWorkflowSecretRepository` becomes a type alias for
  `WorkflowSecretsRepositoryView` (identical surface).
- Doc drift cleanups in route-adapter + create-caller-scope.

Net: ~80 LOC removed without behaviour change.
