+++ packages/platform-api/src/repositories/authorized-workflow-secret-repository.ts
@@ -1,5 +1,4 @@
 import type { CallerIdentity } from '../auth.js';
-import { ForbiddenError } from '../errors.js';
 import { AuthorizedRepository } from './authorized-repository.js';
 
 /**
@@ -24,7 +23,7 @@ export interface WorkflowSecretsRepositoryView {
 }
 
 /** Workspace-scoped workflow secret access. Membership is on the workspace. */
-export interface AuthorizedWorkflowSecretRepository extends WorkflowSecretsRepositoryView {}
+export type AuthorizedWorkflowSecretRepository = WorkflowSecretsRepositoryView;
 
 export class AuthorizedWorkflowSecretRepositoryImpl
   extends AuthorizedRepository<never>
@@ -52,12 +51,12 @@ export class AuthorizedWorkflowSecretRepositoryImpl
     workflowName: string,
     secrets: Record<string, string>,
   ): Promise<void> => {
-    this.assertWrite(namespace);
+    this.assertNamespaceWrite(namespace);
     await this.raw.setSecrets(namespace, workflowName, secrets);
   };
 
   deleteSecrets = async (namespace: string, workflowName: string): Promise<void> => {
-    this.assertWrite(namespace);
+    this.assertNamespaceWrite(namespace);
     await this.raw.deleteSecrets(namespace, workflowName);
   };
 
@@ -67,12 +66,7 @@ export class AuthorizedWorkflowSecretRepositoryImpl
     key: string,
     value: string,
   ): Promise<void> => {
-    this.assertWrite(namespace);
+    this.assertNamespaceWrite(namespace);
     await this.raw.upsertSecret(namespace, workflowName, key, value);
   };
-
-  private assertWrite(namespace: string): void {
-    if (this.caller.kind === 'apiKey') return;
-    if (!this.caller.namespaces.has(namespace)) throw new ForbiddenError();
-  }
 }
